requirementst.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bitarray==2.7.4
certifi==2023.5.7
charset-normalizer==3.1.0
cytoolz==0.12.1
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.1
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
jsonschema==4.17.3
lru-dict==1.2.0
multidict==6.0.4
parsimonious==0.9.0
protobuf==4.23.2
py-solc==3.2.0
py-solc-x==1.1.1
pycryptodome==3.18.0
pyrsistent==0.19.3
python-dotenv==1.0.0
regex==2023.6.3
requests==2.31.0
rlp==3.0.0
semantic-version==2.10.0
toolz==0.12.0
urllib3==2.0.3
web3==6.4.0
websockets==11.0.3
yarl==1.9.2
